+++ src/CF_520A_Panagram.java
@@ -2,7 +2,6 @@
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
-import java.util.ArrayList;
 
 public class CF_520A_Panagram {
 public static void main(String []args) throws Exception{
@@ -17,7 +16,7 @@ public static void main(String []args) throws Exception{
 boolean [] list = new boolean [26];
 for(int i =0 ; i < n ;i++){
 	char g = Character.toLowerCase(y.charAt(i));
-	int ascii =(int) g ;
+	int ascii =g ;
 	int pos = ascii - 97;
 	list[pos]=true;
 }
