+++ src/HR_CamelCase.java
@@ -1,6 +1,6 @@
 
-import java.util.*;
-import java.io.*;
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
 public class HR_CamelCase {
 public static void main(String[]args)throws Exception{
 	BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
@@ -8,7 +8,7 @@ public static void main(String[]args)throws Exception{
 	int answer = 1 ;
 	
 	for(int i =1 ; i<arr.length;i++){
-	int value = (int)arr[i];
+	int value = arr[i];
 	
 	if(value>=65&&value<=90)
 		answer++;
