+++ src/UVA_4524_InterstarTransport.java
@@ -2,9 +2,6 @@
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.PriorityQueue;
 import java.util.StringTokenizer;
 
 
@@ -53,7 +50,7 @@ public static void main(String[]args)throws Exception{
 	adjMatrix = new int [V][V];
 	for(int i =0 ;i<V;i++){
 		for (int j =0 ;j<V;j++){
-			adjMatrix[i][j] = (int)1000000000;
+			adjMatrix[i][j] = 1000000000;
 		}
 	}
 	for(int i=0 ; i<p ; i++){
